create taxes module, export/import toggleTaxes, update functions and references
